refactor issue sync to init to keep main clean.
